Add HostPath example
For someone trying to get started with CSI the existing documentation
stops right at the point where someone is ready to deploy a CSI driver
and experiment with it.

The new example is based on the .yaml files provided by Serguei
Bezverkhi on slack. The troubleshooting section addresses some
real-world issues encountered when trying to get CSI working (local
error with incomplete cluster config and out-dated hostpath image).