example: pull from upstream repos directly, used fixed revisions
Maintaining local copies of all the necessary files is unnecessary, we
might as well tell kubectl to pull directly from the individual
repos. But to ensure that this really pulls the right revision of each
file, we have to specify the revision hashes instead of the "master"
branch.

The markup was changed for these reasons:
- the "copy to clipboard" feature wasn't very useful for the commands
  because it copied the entire command and its output
- the long commands weren't wrapped and the clipboard icon was shown
  on top of them
- inlining the yaml file source into the html version isn't easy
  anymore (would require mdbook preprocessor?), so instead
  all "kubectl create -f" URLs are links to the non-raw version in
  GitHub (non-raw because then one has easy access to commit logs)
- commands and output are now using a different background, which is
  a bit easier to tell them apart

Updating those hashes manually can be cumbersome. The new
hack/update-example.sh script automates that step. The remaining local
.yaml files are the ones that have no corresponding upstream file. It
might be worth pushing them upstream in the future, because then
everything will be in one place (hostpath deployment + examples)
and/or easier to update (first update in upstream repo, then update
docs with the hashes).